requirements.txt
fastapi==0.103.1         # Web framework
uvicorn==0.23.2         # ASGI server
psycopg2-binary==2.9.9  # PostgreSQL connector
faiss-cpu==1.7.4        # FAISS for fast vector search
pgvector==0.2.2         # PostgreSQL vector extension
sentence-transformers==2.2.2  # Pretrained text embeddings
numpy==1.24.3           # Required for cosine similarity
